reference_implementations/gcp/architectures/online/ml-api/requirements.txt
fastapi==0.109.1
fastapi-cli==0.0.4
google-cloud-aiplatform[prediction]==1.66.0
pandas==2.0.3

grpcio-status==1.62.2
numpy<2
streamlit
uvicorn
requests
google-cloud-storage
google-cloud-documentai
